bugfix cost_kgs

+++ flight/models.py
@@ -157,7 +157,7 @@ class BaseParcel(TimeStampedModel):
         code = COD128(f'{self.track_code}', writer=ImageWriter()).write(rv)
         self.barcode.save(f'{self.track_code}.png', File(rv), save=False)
         currency = self.box.destination.currency
-        self.cost_kgs = float(self.cost_usd) * float(currency)
+        self.cost_kgs = int(float(self.cost_usd) * float(currency))
         super(BaseParcel, self).save(*args, **kwargs)
